requirements.txt

matplotlib
numpy
tqdm
box2d
pyglet
gym
h5py
gin-config
tensorflow-gpu==1.12.0
torch
torchvision
